#13 - Feature: 주문 상세 페이지에 예치금 적용

+++ 2Week_Mission/src/main/java/com/ll/exam/ebooks/app/cash/entity/CashLog.java
@@ -25,7 +25,7 @@ public class CashLog extends BaseEntity {
     @ManyToOne(fetch = LAZY)
     private Member member;
 
-    private int price; // 변동 사유
+    private long price; // 변동 사유
 
     private String eventType; // 변동 사유
